add memory-mapped IO
periph/spi_slave is broken; I've gone through a few iterations, and this one was
the closest to working.

periph/uart is a bit heavier (and slower) than it should be.

The interface is quite nice though--a single memory write to 0x800 results in
that word being transmitted at 9600 8n1.