word_encdec: implement encoding for sub-word writes
While experimenting with the behavior of sh/lh, I realized the previous encoder
was massively incomplete--this was partially caused by misunderstanding (or
possibly vagueness in) the spec, and partially caused by not thinking.

The spec says a SH/SB writes the lower bytes of rs2 to the given address. The
old encoder was simply selecting the correct size, but wdata needs to be shifted
to the correct sub-word location based on the address (and the writeb lines
shifted as well).