csr_reg: set mcause and mtval
I decided to implement these writes sequentially with a 2-bit state machine.

This also fixes a minor bug where the trap signal was visible during a pipe
flush--because jump.v doesn't generate trap signals during pipe flush, this was
possible only on the first clock cycle, and had the effect of the first
instruction not being retired.