branch_predict: replace with branch table buffer
This is a (fully parameterized) 4-entry 1-bit btb. The btb is quite
expensive (and can't be implemented using bram without adding more pipeline
stages), but valuable for multiple reasons:

The old/deleted branch_predict.v had a 32-bit adder in the critical path.

Before this commit:

ice40 30.89 MHz ; 2982 LC
ecp5  63.15 MHz ; 3094 LUT4 (2528 SLICE)

After this commit:

ice40 37.35 MHz ; 3500 LC
ecp5  63.53 MHz ; 3769 LUT4 (2444 SLICE)

In addition, branch_predict.v was unable to predict any jalr instructions (due
to being dependent on values from the integer registers, which are not available
in the fetch stage). In the btb implementation, jalr is becomes predictable.
In this implementation, it is useful when jalr is used as a jump table.