reduce alu critical path
alu_zero was in the critical path. This actually never made sense--equality with
zero is just as expensive as a comparison to another number. This adds a new alu
operation, ALU_EQ.

After this commit:

ice40 43.00 MHz
ecp5  71.25 MHz

This also updates alu and jump control logic to match the new alu operations