improve instruction misalign trap handling
The wording of the spec suggests instruction misaligned traps are generated
specifically exclusively by branch instructions, during that instruction--the
branch instruction is not retired.

This makes more sense than the behavior in the previous commit--after this
commit, mepc will be the instruction of the misaligned branch.

This also allows a small optimization: we don't need to add the lower bits of
imm and base (currently in the critical path).