forwarding bug: classify RD_SRC_CSR as a data hazard
While this isn't real a data hazard, because csr reads happen in the execute
stage, in order to make this available via forwarding I would need to either mux
csr_rdata with alu_y in the execute stage on write (undesirable because alu_y is
critical path) or with rs2_rdata in the execute stage on read (undesirable
because it would duplicate logic with the writeback stage, and the forwarding
mux itself is in the critical path).

Because csr instructions don't need to be fast, bubbling for csr data hazards
allows forwarding via the writeback mux.