branch prediction pipelining bug: data_hazard vs prediction pc
A data hazard causes an instruction to be repeated in the decode stage, but by
the time the data hazard is detected, the pc has already been incremented in the
fetch stage.

This is a problem for prediction signaling, because the forwarded prediction
signal during a data hazard would be for the wrong (following) pc value.

This situation happens with jump tables, such as:

  lw t1,0(t0)
  jalr ra,0(t1)

After this commit, the cpu executes my aoc2020 day12 part1 solution correctly.